プレイヤー2のモーション実装＆吹き出し実装

+++ code/_RNLib/Basis/3DObject/Body/doll3D.cpp
@@ -330,7 +330,7 @@ void CDoll3D::DrawModelVtxIdx(CModel::Vertex3DInfo*& vtxInfo, UInt& vtxNum) {
 
 			RNLib::Polygon3D().Put((UShort)RNMode::PRIORITY::UI3D, vtxInfo[cntVtx].worldPos, INITROT3D)
 				->SetSize(0.25f, 0.25f)
-				->SetCol(Color{ 255,0,0,(UShort)(255 * ((float)(drawVtxIdxNum - cntVtx) / drawVtxIdxNum)) })
+				->SetCol(Color{ 255,0,0,(int)(255 * ((float)(drawVtxIdxNum - cntVtx) / drawVtxIdxNum)) })
 				->SetLighting(false)
 				->SetZTest(false)
 				->SetBillboard(true);
@@ -339,7 +339,7 @@ void CDoll3D::DrawModelVtxIdx(CModel::Vertex3DInfo*& vtxInfo, UInt& vtxNum) {
 				->SetLighting(false)
 				->SetZTest(false)
 				->SetBillboard(true)
-				->SetCol(Color{ 255,255,255,(UShort)(255 * ((float)(drawVtxIdxNum - cntVtx) / drawVtxIdxNum)) });
+				->SetCol(Color{ 255,255,255,(int)(255 * ((float)(drawVtxIdxNum - cntVtx) / drawVtxIdxNum)) });
 		}
 	}
 }
